extract Reflector to seperate file
